Move AdminOrderManageTitleNavBarChildren into components/order
